Renamed file from complete_user_registration_contaianer to AfterUserRegistrationContainer
