+++ app/javascript/packs/container/AfterUserRegistrationContainer.vue
