+++ app/javascript/packs/router/router.ts
@@ -3,7 +3,7 @@ import VueRouter from 'vue-router'
 import ArticlesContainer from '../container/ArticlesContainer.vue'
 import LoginContainer from '../container/login_container.vue'
 import UserRegistrationContainer from '../container/user_registration_container.vue'
-import CompleteUserRegistrationContainer from '../container/complete_user_registration_container.vue'
+import CompleteUserRegistrationContainer from '../container/AfterUserRegistrationContainer.vue'
 import PostArticlesContainer from '../container/post_articles_container.vue'
 import MyPageContainer from '../container/mypage_container.vue'
